Update dependencies
Updated to a major new version to the `slack` module. It looks like the RTM handling was dropped from the module altogether, so I re-implemented the communication using the `ws` module. The good news is, there's a ton of different websocket modules for node, so now we just have to find one that automatically reconnects. Changing the websocket implementation should be trivial.